% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bca_mn_priors.R
\name{bca_mn_priors}
\alias{bca_mn_priors}
\title{\code{bvartools} calculates priors slightly differently than
the BCA paper.  This recreates the priors from BCA.}
\usage{
bca_mn_priors(data, lags, gamma1, gamma2, gamma3, gamma4)
}
\arguments{
\item{data, }{matrix}

\item{lags}{number of VAR lags}

\item{gamma1}{number paramater}

\item{gamma2}{number paramater}

\item{gamma3}{number paramater}

\item{gamma4}{number paramater}
}
\value{
list of minneapolis priors
}
\description{
\code{bvartools} calculates priors slightly differently than
the BCA paper.  This recreates the priors from BCA.
}
