% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_vfci.R
\name{est_vfci}
\alias{est_vfci}
\title{Wrapper function to estimate VFCI with a variety of options}
\usage{
est_vfci(
  y,
  x,
  het = NULL,
  vfci_vars = NULL,
  date_col = "date",
  forward = 1,
  lags = 0,
  cols_to_lag = y,
  exlags = FALSE,
  date_begin = "1962-04-01",
  date_end = "2022-07-01",
  include_all = FALSE,
  data = fread("./data/vfci_data.csv")
)
}
\arguments{
\item{y}{dependent variable, assumed to be given in log levels,
will take forward differences}

\item{x}{independent variables, usually the financial principal components,
x = c("pc1", "pc2", "pc3", "pc4")}

\item{het}{defaults to the same as x, but can be a subset of columns or other columns}

\item{vfci_vars}{defaults to the same as het, but can be a subset of columns or other columns}

\item{date_col}{which column in the data.table identifies the 'date' or time domain}

\item{forward}{how many periods to take forward differences of y}

\item{lags}{how many lags to include in the independent variables, defaults to 0}

\item{cols_to_lag}{which columns to lags of, defaults to just 'y'}

\item{exlags}{boolean, should lags be included in the vfci calculation or excluded}

\item{date_begin}{start date for the data}

\item{date_end}{end date for the data}

\item{include_all}{boolean, whether to return all the get_vfci() data
(ie model estimation) or just the time series data}

\item{data}{A data.table object containing data for the VFCI estimation,
defaults to vfci_data.csv file}
}
\value{
data.table object with 'date', 'vfci', and 'mu' columns
}
\description{
Wrapper function to estimate VFCI with a variety of options
}
