% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bvartools.R
\name{estimate_bvartools}
\alias{estimate_bvartools}
\title{Wrapper function to create a \code{bvartools} estimation
Takes priors and how many itertions to run, and
runs a gibbs sampler to find draws of coefficients
and the sigma matrix.}
\usage{
estimate_bvartools(v, priors, iterations = 1000, burnin = 1000)
}
\arguments{
\item{v}{var from gen_var()}

\item{priors}{priors from bca_mn_priors()}

\item{iterations}{number of iterations to run, default to 1000}

\item{burnin}{extra iterations to "burn in" at the beginning, default to 1000}
}
\value{
bvartools bvar
}
\description{
Wrapper function to create a \code{bvartools} estimation
Takes priors and how many itertions to run, and
runs a gibbs sampler to find draws of coefficients
and the sigma matrix.
}
