% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_linear_het_reg.r
\name{find_linear_het_reg_q_mriv}
\alias{find_linear_het_reg_q_mriv}
\title{Return a single column vector of the Q rotation matrix.
Uses the IV method in Mertens Rvan (2013) Appendix A.}
\usage{
find_linear_het_reg_q_mriv(var, het_reg, target)
}
\arguments{
\item{var}{a VAR}

\item{het_reg}{het_reg from fit_het_reg_from_var}

\item{target}{variable in VAR use to identify Q column}
}
\value{
A single column vector of the orthonormal matrix Q
}
\description{
Return a single column vector of the Q rotation matrix.
Uses the IV method in Mertens Rvan (2013) Appendix A.
}
