% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_from_var.r
\name{get_data_from_var}
\alias{get_data_from_var}
\title{Recover data from a vars::VAR object with any number of lags.
Note that date columns are not saved in a var and cannot be recovered.}
\usage{
get_data_from_var(var)
}
\arguments{
\item{var}{vars::VAR}
}
\value{
A data.frame
}
\description{
Recover data from a vars::VAR object with any number of lags.
Note that date columns are not saved in a var and cannot be recovered.
}
