% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recession_dt.R
\name{get_recession_dt}
\alias{get_recession_dt}
\title{Convert recession_dates.csv into a data.table of start and end dates.}
\usage{
get_recession_dt(path = "./data-raw/recession_dates.csv")
}
\arguments{
\item{path}{filepath to the recession_dates.csv}
}
\value{
data.table of recession start and end dates
}
\description{
Convert recession_dates.csv into a data.table of start and end dates.
}
