% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vfci.R
\name{get_vfci}
\alias{get_vfci}
\title{Wrapper function to estimate VFCI with a variety of options}
\usage{
get_vfci(
  data,
  y,
  x,
  het = x,
  vfci_vars = het,
  date_col = "date",
  method = "ML",
  ...,
  gls_opt = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{y}{dependent variable}

\item{x}{independent variables}

\item{het}{heteroskedasticity variables}

\item{vfci_vars}{variables to use to calculate VFCI, can be a subset of het}

\item{date_col}{date column}

\item{method}{"ML" or "twostep", whether to use Maximum Likelihood or two step
estimation of the heteroskedastic regression.}

\item{...}{additional arguments to pass to hetreg}

\item{gls_opt}{additional arguments to pass to gls}
}
\value{
list of VFCI ts, hetreg model, and call
}
\description{
Wrapper function to estimate VFCI with a variety of options
}
