% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetreg_lrtest.R
\name{hetreg_lrtest}
\alias{hetreg_lrtest}
\title{Function to calculate liklihood ratio test for hetreg}
\usage{
hetreg_lrtest(hetreg)
}
\arguments{
\item{hetreg}{a hetreg class object, method must be from \code{hetreg_gls}}
}
\value{
a \code{hetreg_waldtest} object
}
\description{
Function to calculate liklihood ratio test for hetreg
}
