% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetreg_twostep.R
\name{hetreg_twostep}
\alias{hetreg_twostep}
\title{Fits a heteroskedastic regression by twostep method}
\usage{
hetreg_twostep(data, y, x1, x2 = x1)
}
\arguments{
\item{data}{dataframe}

\item{y}{dependent variable}

\item{x1}{independent variables for predicting mean}

\item{x2}{indepedent varaibles for predicting heteroskedasticity,
defaults to x1}
}
\value{
a \code{hetreg} object
}
\description{
Fits a heteroskedastic regression by twostep method
}
