% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetreg_waldtest.R
\name{hetreg_waldtest}
\alias{hetreg_waldtest}
\title{Function to calculate waldtest for hetreg object to test that all
coefficients in the heteroskedastic regression are jointly non-zero.
Code based on https://github.com/cran/aod/blob/master/R/stat.test.R}
\usage{
hetreg_waldtest(hetreg)
}
\arguments{
\item{hetreg}{a hetreg class object, method must be "twostep"}
}
\value{
a \code{hetreg_waldtest} object
}
\description{
Function to calculate waldtest for hetreg object to test that all
coefficients in the heteroskedastic regression are jointly non-zero.
Code based on https://github.com/cran/aod/blob/master/R/stat.test.R
}
