% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_corr.r
\name{hs_corr}
\alias{hs_corr}
\title{Calculate the correlation between the structural shocks of a list of VARs.}
\usage{
hs_corr(var_list)
}
\arguments{
\item{var_list}{a list of vars::VAR}
}
\value{
A data.frame
}
\description{
Calculate the correlation between the structural shocks of a list of VARs.
}
