% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\alias{print.hetreg}
\alias{print.hetreg_waldtest}
\alias{print.hetreg_lrtest}
\title{Print a hetreg object}
\usage{
\method{print}{hetreg}(x, ...)

\method{print}{hetreg_waldtest}(x, digits = 4, ...)

\method{print}{hetreg_lrtest}(x, digits = 4, ...)
}
\arguments{
\item{x}{hetreg_lrtest object}

\item{...}{NOT USED}

\item{digits}{number of digits to display, defaults to 4}
}
\description{
Print a hetreg object

Print a hetreg waldtest

Print a hetreg LR test
}
