% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_var.r
\name{reorder_var}
\alias{reorder_var}
\title{Reestimate a VAR with the given data columns first.
Can take indexes or column names.
Can reorder 1, 2, ..., or all columns.}
\usage{
reorder_var(var, new_order)
}
\arguments{
\item{var}{vars::VAR, original var}

\item{new_order}{column names or indexes, columns to be moved to front}
}
\value{
A vars::VAR object
}
\description{
Reestimate a VAR with the given data columns first.
Can take indexes or column names.
Can reorder 1, 2, ..., or all columns.
}
