% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.hetreg}
\title{Nicely summarize a hetreg object}
\usage{
\method{summary}{hetreg}(object, ...)
}
\arguments{
\item{object}{hetreg object}

\item{...}{NOT USED}
}
\description{
Nicely summarize a hetreg object
}
