% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_paper.r
\docType{data}
\name{theme_paper}
\alias{theme_paper}
\title{{ggplot2} theme to use for paper figures.}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 136.
}
\usage{
theme_paper
}
\description{
{ggplot2} theme to use for paper figures.
}
\keyword{datasets}
