% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_direct.R
\name{var_direct}
\alias{var_direct}
\title{Directly estimate a vars::VAR object using \code{y} and \code{x} rather than \code{y} and some number of lags.}
\usage{
var_direct(y, x, type = c("const", "trend", "both", "none"), ...)
}
\arguments{
\item{y}{endogenous variables}

\item{x}{exogenous variables}

\item{type}{a character specifying if a constant or trend is included in the model.}

\item{...}{NOT USED}
}
\value{
a list with class attrribute \code{vars::varest} holding the following elements:
\itemize{
\item \code{varresult}: A list of \code{lm} objects, one per equation.
\item \code{datamat}: The data matrix of endogenous and exogenous variables.
\item \code{type}: A character specifying if a constant or trend is included in the model.
\item \code{p}: an integer specifying the lag order (set to ncol(x) mod ncol(y)).
\item \code{K}: An integer specifying the dimension of the VAR.
\item \code{nobs}: An integer specifying the number of observations.
\item \code{totobs}: An integer specifying the total number of observations.
\item \code{restrictions}: NULL
\item call: the \code{call} to \code{VAR_direct()}
}
}
\description{
Estimates a VAR by OLS per equation.
}
