% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_labels.r
\docType{data}
\name{variable_labels}
\alias{variable_labels}
\title{Labels for the column names.  Useful when making plots.}
\format{
An object of class \code{character} of length 11.
}
\usage{
variable_labels
}
\description{
Labels for the column names.  Useful when making plots.
}
\keyword{datasets}
