% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_var.r
\name{fit_var}
\alias{fit_var}
\title{Fit a VAR to given data with given lags. Ignores a column called "date" and includes a constant.}
\usage{
fit_var(data, lags, date_col = "date", include_const = TRUE)
}
\arguments{
\item{data}{data.frame, with a date column}

\item{lags}{integer, number of lags to use in VAR}

\item{date_col}{string, defaults to "date", name of the date column}

\item{include_const}{boolean, TRUE (default) includes a constant in the VAR estimation}
}
\value{
A vars::VAR object
}
\description{
Fit a VAR to given data with given lags. Ignores a column called "date" and includes a constant.
}
