% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_var_data.r
\name{get_var_data}
\alias{get_var_data}
\title{Function to subset all_analysis_data to a data.table with requested variables.
Defaults to the ones used in BCA + vfci.}
\usage{
get_var_data(
  data_path = "./data/all_analysis_data.rds",
  output = "output",
  investment = "investment",
  consumption = "consumption",
  hours_worked = "hours_worked",
  unemployment = "unemployment",
  labor_share = "labor_share",
  interest = "interest",
  inflation = "inflation",
  productivity = "productivity",
  TFP = "TFP",
  vfci = "vfci_fgr10gdpc1",
  start_date = as.Date("1962-01-01"),
  end_date = as.Date("2017-01-01"),
  cols = NULL,
  add_cols = NULL
)
}
\arguments{
\item{data_path}{string, path to data file to read in}

\item{output}{string, column name to be labelled output}

\item{investment}{string, column name to be labelled investment}

\item{consumption}{string, column name to be labelled consumption}

\item{hours_worked}{string, column name to be labelled hours_worked}

\item{unemployment}{string, column name to be labelled unemployment}

\item{labor_share}{string, column name to be labelled labor_share}

\item{interest}{string, column name to be labelled interest}

\item{inflation}{string, column name to be labelled inflation}

\item{productivity}{string, column name to be labelled productivity}

\item{TFP}{string, columne name to be labelled TFP}

\item{vfci}{string, column name to be labelled vfci, defaults to vfci_fgr10gdpc1}

\item{start_date}{date, start date of data}

\item{end_date}{date, end date of data}

\item{cols}{string vector of columns to subset, names will be used as new columns names}

\item{add_cols}{sting vector of columns to include in the subest with the default ones}
}
\value{
data.table
}
\description{
Function to subset all_analysis_data to a data.table with requested variables.
Defaults to the ones used in BCA + vfci.
}
