% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_linear_het_reg.r
\name{id_linear_het_reg}
\alias{id_linear_het_reg}
\title{Use the fit of a linear regression estimating the heteroskedasticity
of the variance of the residuals in a VAR to identify a structural VAR.}
\usage{
id_linear_het_reg(
  var,
  target,
  het_reg_lags = 0,
  constant = TRUE,
  use_hetreg_func = TRUE,
  sign = "pos",
  sign_horizon = 1,
  method = "default"
)
}
\arguments{
\item{var}{a VAR, either vars::VAR or svars object}

\item{target}{variable in VAR to target with het_reg}

\item{het_reg_lags}{lags passed to het_reg}

\item{constant, }{boolean, default to TRUE for constant in het_reg regression}

\item{use_hetreg_func, }{boolean, default to TRUE to use hetreg function}

\item{sign}{Default to "positive". Can be "negative".  Ensures the
cummulative impact of the main shock on the target variable is the
given sign.}

\item{sign_horizon}{Default to 1. The horizon through which to accumulate the
impact of the shock.}

\item{method, }{"default" for our derivations for finding the q column,
and "mriv" to use Mertens and Ravn (2013) Appendix A IV derivations.}
}
\value{
SVAR
}
\description{
Use the fit of a linear regression estimating the heteroskedasticity
of the variance of the residuals in a VAR to identify a structural VAR.
}
